refactor: Update import paths for otpRoutes and authRoutes
The code changes in `index.ts` update the import paths for `otpRoutes` and `authRoutes` to reflect the new directory structure. This ensures that the correct files are imported and used in the application.
